Start moving directories, plan file, and extension to Engine.
This is turning out to be too big and too messy, so I'm going to throw it out
for now. But I'm committing in a branch for later reference, as it might be
useful when I have some tuits to work on a reorganization of the code to
better deal with configuration hierarchies.